+++ src/bittytax/conv/parsers/__init__.py
@@ -1,5 +1,6 @@
 from . import (
     accointing,
+    adalite,
     barclays,
     binance,
     bitfinex,