+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Conversion tool: added parser for Subscan explorer. ([#305](https://github.com/BittyTax/BittyTax/issues/305))
 - KuCoin parser: added BillingHistory bundle files. ([#307](https://github.com/BittyTax/BittyTax/issues/307))
 - CEX.IO parser: added additional staking types. ([#297](https://github.com/BittyTax/BittyTax/issues/297))
+- Conversion tool: added parser for AdaLite wallet. ([#221](https://github.com/BittyTax/BittyTax/issues/221))
 ### Removed
 - Conversion tool: removed merge parser for Coinbase/Coinbase Pro.